unsignedBigInteger

+++ database/migrations/2014_10_12_000001_create_profiles_table.php
@@ -15,7 +15,7 @@ class CreateProfilesTable extends Migration
     {
         Schema::create('profiles', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('user_id');
+            $table->unsignedBigInteger('user_id');
             $table->date('dob')->nullable();
             $table->string('cnic')->nullable();
             $table->string('cnic_front')->nullable();
