+++ database/migrations/2020_01_27_034947_create_orders_table.php
@@ -15,9 +15,9 @@ class CreateOrdersTable extends Migration
     {
         Schema::create('orders', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('lifter_id');
-            $table->integer('consumer_id');
-            $table->integer('service_id');
+            $table->unsignedBigInteger('lifter_id');
+            $table->unsignedBigInteger('consumer_id');
+            $table->unsignedBigInteger('service_id');
             $table->integer('qty');
             $table->double('price');
             $table->integer('charges')->default(0);
