+++ database/migrations/2020_02_12_071443_create_open_orders_table.php
@@ -15,8 +15,8 @@ class CreateOpenOrdersTable extends Migration
     {
         Schema::create('open_orders', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('consumer_id');
-            $table->integer('service_id');
+            $table->unsignedBigInteger('consumer_id');
+            $table->unsignedBigInteger('service_id');
             $table->integer('qty');
             $table->double('price');
             $table->integer('charges')->default(0);
