+++ database/migrations/2020_02_24_165024_create_schedule_orders_table.php
@@ -15,9 +15,9 @@ class CreateScheduleOrdersTable extends Migration
     {
         Schema::create('schedule_orders', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('consumer_id');
-            $table->integer('lifter_id');
-            $table->integer('service_id');
+            $table->unsignedBigInteger('consumer_id');
+            $table->unsignedBigInteger('lifter_id');
+            $table->unsignedBigInteger('service_id');
             $table->integer('qty');
             $table->double('price');
             $table->integer('charges');
